+++ lib/im_localized.dart
@@ -4,3 +4,5 @@ export 'src/parsing/translations.dart';
 export 'src/widget/im_localized_app.dart';
 export 'src/widget/im_localized_controller.dart';
 export 'src/utils/extensions.dart';
+export 'src/storage/localized_storage.dart';
+export 'src/storage/shared_preferences_storage.dart';