+++ lib/src/utils/select_locale.dart
@@ -3,11 +3,14 @@ import 'package:im_localized/im_localized.dart';
 
 Locale selectLocale(
   List<Locale> supportedLocales,
-  Locale deviceLocale, {
-  Locale? fallbackLocale,
+  Locale desiredLocale, {
+  List<Locale> fallbackLocales = const [],
 }) {
   return supportedLocales.firstWhere(
-    (locale) => locale.supports(deviceLocale),
-    orElse: () => fallbackLocale ?? supportedLocales.first,
+    (locale) => locale.supports(desiredLocale),
+    orElse: () => fallbackLocales.firstWhere(
+      (fallback) => fallback.supports(desiredLocale),
+      orElse: () => supportedLocales.first,
+    ),
   );
 }