+++ lib/src/widget/im_localized_controller.dart
@@ -3,91 +3,158 @@ import 'package:im_localized/im_localized.dart';
 import 'package:im_localized/src/utils/select_locale.dart';
 import 'package:intl/intl_standalone.dart'
     if (dart.library.html) 'package:intl/intl_browser.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:im_localized/src/utils/utils.dart';
 
 class ImLocalizedController extends ChangeNotifier {
-  static Locale? _savedLocale;
-  static late Locale _deviceLocale;
+  final List<Locale> fallbackLocales;
 
-  late Locale _locale;
+  final LocaleStorage? localeStorage;
 
-  final bool useFallbackTranslations;
-  final bool saveLocale;
-  Translations? _translations;
-  Translations? _fallbackTranslations;
-  Translations? get translations => _translations;
-  Translations? get fallbackTranslations => _fallbackTranslations;
+  final TranslationsStorage? translationsStorage;
 
   ImLocalizedController({
     required List<Locale> supportedLocales,
-    required this.useFallbackTranslations,
-    required this.saveLocale,
+    this.localeStorage,
+    this.translationsStorage,
     Translations? translations,
     Locale? startLocale,
-    Locale? fallbackLocale,
-    Locale? forceLocale, // used for testing
+    this.fallbackLocales = const [],
   }) {
-    if (saveLocale && _savedLocale != null) {
-      ImLocalizedApp.logger.d('Saved locale loaded ${_savedLocale.toString()}');
-      _locale = selectLocale(
-        supportedLocales,
-        _savedLocale!,
-        fallbackLocale: fallbackLocale,
-      );
-    } else {
-      // From Device Locale
-      _locale = selectLocale(
-        supportedLocales,
-        _deviceLocale,
-        fallbackLocale: fallbackLocale,
-      );
-    }
-
-    _translations = translations != null && translations.activeLocale != _locale
-        ? translations.copyWith(activeLocale: _locale)
-        : _translations;
+    _translations = translations;
+    _init(supportedLocales);
   }
 
-  Locale get locale => _locale;
+  Locale? _savedLocale;
 
-  Future<void> setLocale(Locale nextLocale) async {
+  bool _initializing = false;
+
+  late Locale _deviceLocale;
+
+  Locale? _locale;
+
+  Translations? _translations;
+
+  Translations? _fallbackTranslations;
+
+  //******************************************/
+  //    Getters
+  //******************************************/
+
+  Translations? get translations => _translations;
+
+  Translations? get fallbackTranslations => _fallbackTranslations;
+
+  bool get initializing => _initializing;
+
+  bool get initialized => _locale != null;
+
+  Locale get deviceLocale => _deviceLocale;
+
+  Locale? get savedLocale => _savedLocale;
+
+  Locale get locale => _locale ?? const Locale('und');
+
+  //******************************************/
+  //    Methods
+  //******************************************/
+
+  /// sets new locale and saves it to storage if localeStorage is not null
+  Future<void> setLocale(Locale nextLocale, {saveToStorage = true}) async {
+    if (!initialized) {
+      throw Exception('ImLocalizedController.setLocale: '
+          'Locale not initialized yet');
+    }
+
     _locale = nextLocale;
     _translations = _translations?.copyWith(activeLocale: _locale);
     notifyListeners();
     ImLocalizedApp.logger.d('Locale $locale changed');
-    await _saveLocale(_locale);
+    if (saveToStorage && localeStorage != null) {
+      await localeStorage!.saveLocale(_locale!.toString());
+      ImLocalizedApp.logger.d('Locale $locale saved');
+    }
   }
 
-  Future<void> _saveLocale(Locale? locale) async {
-    if (!saveLocale) return;
-    final preferences = await SharedPreferences.getInstance();
-    await preferences.setString('locale', locale.toString());
-    ImLocalizedApp.logger.d('Locale $locale saved');
-  }
-
-  static Future<void> init() async {
-    final preferences = await SharedPreferences.getInstance();
-    final strLocale = preferences.getString('locale');
-    _savedLocale = strLocale?.toLocale();
-    final foundPlatformLocale = await findSystemLocale();
-    _deviceLocale = foundPlatformLocale.toLocale();
-    ImLocalizedApp.logger.d('Localization initialized');
-  }
-
+  /// removes saved locale from storage, but does not change the current locale
   Future<void> deleteSavedLocale() async {
+    if (!initialized) {
+      throw Exception('ImLocalizedController.deleteSavedLocale: '
+          'Locale not initialized yet');
+    }
+
     _savedLocale = null;
-    final preferences = await SharedPreferences.getInstance();
-    await preferences.remove('locale');
-    ImLocalizedApp.logger.d('Saved locale deleted');
+    if (localeStorage != null) {
+      await localeStorage!.deleteLocale();
+    }
   }
 
-  Locale get deviceLocale => _deviceLocale;
-
+  /// removes saved locale from storage and sets locale to device locale
   Future<void> resetLocale() async {
-    ImLocalizedApp.logger.d('Reset locale to platform locale $_deviceLocale');
+    if (!initialized) {
+      throw Exception('ImLocalizedController.resetLocale: '
+          'Locale not initialized yet');
+    }
 
-    await setLocale(_deviceLocale);
+    ImLocalizedApp.logger.d('Reset locale to platform locale $_deviceLocale');
+    await setLocale(_deviceLocale, saveToStorage: false);
+    if (localeStorage != null) {
+      await localeStorage!.deleteLocale();
+    }
+  }
+
+  /// Sets new translations and saves it to storage if translationsStorage
+  /// is not null
+  Future<void> replaceTranslations(
+    List<Map<LocalizationKey, LocalizationValue>> next,
+  ) async {
+    final nextTranslations =
+        Translations.fromList(next, activeLocale: _translations?.activeLocale);
+
+    _translations = nextTranslations;
+    notifyListeners();
+    if (translationsStorage != null) {
+      return translationsStorage!.saveTranslations(_translations!);
+    }
+  }
+
+  //******************************************/
+  //    private
+  //******************************************/
+
+  Future<void> _init(List<Locale> supportedLocales) async {
+    WidgetsFlutterBinding.ensureInitialized();
+    _initializing = true;
+
+    await Future.wait([
+      // get deviceLocale
+      findSystemLocale().then((value) {
+        _deviceLocale = value.toLocale();
+      }),
+      // get saved locale if localeStorage is not null
+      if (localeStorage != null)
+        localeStorage!.loadLocale().then((localeString) {
+          _savedLocale = localeString?.toLocale();
+        }),
+      // get translations if translationsStorage is not null
+      if (translationsStorage != null)
+        translationsStorage!.loadTranslations().then((translations) {
+          _translations = translations ?? _translations;
+        }),
+    ]);
+
+    _locale = selectLocale(
+      supportedLocales,
+      _savedLocale ?? deviceLocale,
+      fallbackLocales: fallbackLocales,
+    );
+
+    _initializing = false;
+
+    assert(initialized);
+
+    notifyListeners();
+
+    ImLocalizedApp.logger.d('Localization initialized');
   }
 }